requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
soundfile>=0.12.1
rich>=13.7.0
typer>=0.9.0
python-dotenv>=1.0.0
httpx>=0.25.0

# Optional: Voice Activity Detection (saves bandwidth/costs)
# Uncomment to enable (requires C++ compiler on Windows):
# webrtcvad>=2.0.10
